requirements.txt
numpy==1.22.0
opencv-python
PyYAML==6.0.2
tqdm
scipy
soundfile==0.12.1
resampy
requests
librosa
numba

aiohttp
aiohttp_cors
aiortc
av==12.3.0

edge_tts
pyttsx3
funasr==1.0.23
modelscope==1.8.1
onnxruntime-gpu>=1.18.0
-f https://k2-fsa.github.io/sherpa/onnx/cuda.html
sherpa-onnx==1.12.19+cuda12.cudnn9

setuptools<81
